Split tests for ImplementInterface and ExtractInterface into refactoring and refactoring action part
